+++ doc/source/changes.rst
@@ -12,6 +12,9 @@ New in 0.18
   and replaced with a new parameter called ``incomplete``.
 * API: Several plotting functions have a new ``xticklabels`` parameter to
   suppress x-axis tick labels (e.g. :class:`plot.UTSStat`).
+* The objects returned by :mod:`plot.brain` plotting functions now contain
+  a :meth:`~plot._brain_mixin.BrainMixin.plot_colorbar` method to create a
+  corresponding :class:`plot.ColorBar` plot.
 * New function :func:`choose` to combine data in different :class:`NDVars`
   on a case by case basis.
 * Rejection-GUI (:func:`gui.select_epochs`): Press Shift-i when hovering over